Display and jump to Calibre-style references
When a selection is highlighed, use current selection CFI to create
a Calibre Reference Mode-style citation (#.#), based on the spine
item the current selection belongs to.

This PR also adds a Citation tab to the sidebar, where a Calibre-
style reference or CFI (e.g. from a bookmark) can be entered. After
a user presses "Go" or hits enter on their keyboard, the epub will
jump to the correct paragraph. If an invalid paragraph number is
entered, the epub will instead jump to the beginning of the chapter.